Setup all initial files Build the book's object
Use Object Constructor to create a collection of books.
using Constructor, is the recommended way to achieve this, because all
the books has to have the same information such as title, author, pages and
status.

Choose to create a method in the prototype instead of the constructor.
It's the recommended way to create a method for constructor function.
it will avoid create the same method every time we create a new instance of
the object, in this case a new book.

Create a fucntion do create new book and also a function to add new books in the collection (library)